tests/wpt/web-platform-tests/IndexedDB/idbdatabase_createObjectStore3.htm

<!DOCTYPE html>
<title>IDBDatabase.createObjectStore() - attempt to create an object store outside of a version change transaction </title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>

<script>

var t = async_test(),
    open_rq = createdb(t)

open_rq.onupgradeneeded = function() {}
open_rq.onsuccess = function (e) {
    var db = e.target.result
    assert_throws_dom(
        'InvalidStateError',
        function() { db.createObjectStore('fails') })
    t.done();
}

</script>

<div id="log"></div>
